Remove heroic attempt to be binary compatible
On WebpayMallNormal I tried to keep around the old method that took a
specific ArrayList. But it won't work as at least the JSP compiler
doesn't like the situation where both methods are almost identical.

Given that it's highly unlikely that someone will use the transbank
SDK without recompiling (and if it were the case, recompiling will
easily fix the problem), I'm accepting the slight risk of backwards
incompatibility with the legacy libwebpay.