Catch what really should be the last snake_case
They are truly gone now. A thing of the past. Not to be remembered nor
honored. Just something that shouldn't have happened.

PS: Sonar gets a lot more credit for being persistent than me here.

PS2: Don't point me to the webpay example project. There is no way I'll
be connecting Sonar to it. I don't want to spend a whole day or more
fixing the 300+ issues that will be triggered there.