Integrate Allware's PatPass SDK
This is far from being verbatim from the original legacy SDK. I
refrained from copying anyhing that was a duplicate from the Webpay
code and found out that most of it could be reused by changing any
reference to cl.transbank.patpass.* to cl.transbank.webpay.*.

(Why? Because the API is exactly the same as the Webpay Normal flow,
with a flag and some "WPM" details. WPM stands for WebPay Mensual
which is some kind of ancient/internal name for PatPass by Webpay)

I also took some different design decisions given that we don't have
to be backwards compatible here (unlike webpay where there are plenty
of people using the SDK, it's unlikely that people are using the
PatPass SDK)

- Avoid making initTransaction's parameter list too long, by
  introducing the PatPassInfo object to hold the extra information
  needed by PatPass.

- Use a enum to set the "uf flag", so that the code is more
  self-explanatory (instead of a plain boolean).

- Keep using Webpay as the "entry point" that can be used to obtain a
  "PatPass transaction".

- Use the actual product name ("PatPass by Webpay"). I kept the
  "normal" suffix given that almost every product gets additional stuff
  later on. So who knows if we might get a PatPass Mall in the future!

I kept the fact that the UF flag and the merchant email is set on the
Configuration rather than individually on each transaction. It seemed
like a good decision to me, even though it added more stuff to the
webpay.configuration.Configuration class.