Fix incompatibility with some JAX-WS environments.
The legacy code was expecting CXF to be the default JAX-WS provider.
If it wasn't, ClassCastExceptions were thrown when trying to use
CXF-specific features (the interceptors used for signatures and
validations) on the system-provided JAX-WS Services.

The fix is to create the Service instances ourselves using what CXF
provides: the JaxWsProxyFactoryBean. That way we ensure that our
Service instances are indeed CXF instances and therefore we can use
CXF-specific things on them.